* move decorator to individual module and open for different output implementations * move bytes utility to separated module
